+++ Engine/matrix.cpp
@@ -18,7 +18,7 @@ void Matrix::SetIdentity() {
     }
 }
 
-void Matrix::SetRotation(float x, float y, float z) {
+void Matrix::SetRotation(double x, double y, double z) {
 	SetIdentity();
 	data[0][0] = cos(y)*cos(z);
 	data[0][1] = cos(x)*sin(z) + sin(x)*sin(y)*cos(z);
@@ -31,7 +31,7 @@ void Matrix::SetRotation(float x, float y, float z) {
 	data[2][2] = cos(x)*cos(y);
 }
 
-void Matrix::SetRotation(float theta, Vector3D u) {
+void Matrix::SetRotation(double theta, Vector3D u) {
 	SetIdentity();
 	data[0][0] = cos(theta) + u.x*u.x*(1-cos(theta));
 	data[0][1] = u.x*u.y*(1-cos(theta)) - u.z*sin(theta);
@@ -91,7 +91,7 @@ bool Matrix::operator!= (const Matrix& rhs) const {
     return true;  
 }
    
-Matrix Matrix::operator* (const float rhs) const {
+Matrix Matrix::operator* (const double rhs) const {
     Matrix m(*this);
     for (int row=0;row<4;row++) {
         for (int col=0;col<4;col++) {
@@ -101,7 +101,7 @@ Matrix Matrix::operator* (const float rhs) const {
     return m;
 }
     
-Matrix Matrix::operator/ (const float rhs) const {
+Matrix Matrix::operator/ (const double rhs) const {
     Matrix m(*this);
     for (int row=0;row<4;row++) {
         for (int col=0;col<4;col++) {
