+++ Engine/object.cpp
@@ -14,7 +14,7 @@ Object::Object()
 }
 
 //Alternate texture, veloctity, heading, position
-Object::Object(const unsigned short* texture, const int texWid, const int texHgt, Vector3D vel, Point3D pos, float _rx, float _ry, float _rz)
+Object::Object(const unsigned short* texture, const int texWid, const int texHgt, Vector3D vel, Point3D pos, double _rx, double _ry, double _rz)
 {
 	velocity = vel;
 	position = pos;
@@ -28,7 +28,7 @@ Object::Object(const unsigned short* texture, const int texWid, const int texHgt
 
 //Alternate Constructor
 Object::Object(Poly poly, const unsigned short* texture, const int texWid, const int texHgt, Vector3D vel, 
-		Point3D pos, float _rx, float _ry, float _rz)
+		Point3D pos, double _rx, double _ry, double _rz)
 {
 	velocity = vel;
 	position = pos;
@@ -105,8 +105,8 @@ void Object::Translate(Vector3D& v) {
 }
 
 void Object::RotateToHeading() {
-	float r = cos(phi);
-	float x = r*sin(theta),
+	double r = cos(phi);
+	double x = r*sin(theta),
 			y = sin(phi),
 			z = r*cos(theta);
 	Vector3D forward = normalize(Vector3D(x,y,z));
@@ -127,8 +127,8 @@ void Object::RotateToHeading() {
 }
 
 void Object::RotateToHeading(Vector3D changeUp) {
-	float r = cos(phi);
-	float x = r*sin(theta),
+	double r = cos(phi);
+	double x = r*sin(theta),
 			y = sin(phi),
 			z = r*cos(theta);
 	Vector3D forward = normalize(Vector3D(x,y,z));
@@ -168,8 +168,8 @@ void Object::TranslateTemp(const Vector3D& v) {
 }
 
 void Object::RotateTempToHeading() {
-	float r = cos(phi);
-	float x = r*sin(theta),
+	double r = cos(phi);
+	double x = r*sin(theta),
 			y = sin(phi),
 			z = r*cos(theta);
 	Vector3D forward = normalize(Vector3D(x,y,z));
@@ -194,7 +194,7 @@ void Object::add(Poly poly) {
 	updateList();
 }
 
-void Object::GenerateCube(float size) {	  
+void Object::GenerateCube(double size) {	  
 	master.clear();
 
 	//stores in objects master list
@@ -237,11 +237,11 @@ void Object::GenerateCube(float size) {
 	for(std::list<Poly>::iterator it = master.begin(); it != master.end(); ++it)
 	{
 		it->SetTexture(tex, texWidth, texHeight);
-		it->SetColor(rand()/(float)RAND_MAX,rand()/(float)RAND_MAX,rand()/(float)RAND_MAX);
+		it->SetColor(rand()/(double)RAND_MAX,rand()/(double)RAND_MAX,rand()/(double)RAND_MAX);
 #if 0
-		it->SetVertexColors(rand()/(float)RAND_MAX,rand()/(float)RAND_MAX,rand()/(float)RAND_MAX,
-			rand()/(float)RAND_MAX,rand()/(float)RAND_MAX,rand()/(float)RAND_MAX,
-			rand()/(float)RAND_MAX,rand()/(float)RAND_MAX,rand()/(float)RAND_MAX);
+		it->SetVertexColors(rand()/(double)RAND_MAX,rand()/(double)RAND_MAX,rand()/(double)RAND_MAX,
+			rand()/(double)RAND_MAX,rand()/(double)RAND_MAX,rand()/(double)RAND_MAX,
+			rand()/(double)RAND_MAX,rand()/(double)RAND_MAX,rand()/(double)RAND_MAX);
 #else
 		it->SetVertexColors(0.9,0,0,
 			0,0.9,0,
@@ -258,7 +258,7 @@ void Object::GenerateCube(float size) {
 	updateList(); 
 }
 
-void Object::GenerateTetra(float size) {
+void Object::GenerateTetra(double size) {
 	Vertex p1 = Vertex(size,0,-size/sqrt(2.0)),
 			p2 = Vertex(-size,0,-size/sqrt(2.0)),
 			p3 = Vertex(0,size,size/sqrt(2.0)),
@@ -288,7 +288,7 @@ void Object::GenerateTetra(float size) {
 	updateList(); 
 }
 
-void Object::GenerateFloor(float length, float depth) {
+void Object::GenerateFloor(double length, double depth) {
 
 	master.clear();
 	master.push_back( Poly( Vertex(-length,0,-length,1,0,1),Vertex(-length,0,length,1,0,0),Vertex(length,0,length,1,1,0),
@@ -311,7 +311,7 @@ void Object::GenerateFloor(float length, float depth) {
 }
 
 //Incomplete...
-void Object::GenerateCeiling(float length, float depth) {
+void Object::GenerateCeiling(double length, double depth) {
 	position = Point3D(0,depth,0);
 
 	theta = 0;
@@ -341,7 +341,7 @@ void Object::GenerateCeiling(float length, float depth) {
 ///////////// 1 = left: the left of player init
 ///////////// 2 = right: to the right of player init
 ///////////// 3 = back: behind player init
-void Object::GenerateWall(size_t type, float length, float depth) {
+void Object::GenerateWall(size_t type, double length, double depth) {
 	
 	switch (type)
 	{
@@ -389,7 +389,7 @@ void Object::GenerateWall(size_t type, float length, float depth) {
 	updateList(); 
 }
 
-void Object::GenerateShot(Vector3D pos, float theta_, float phi_) {
+void Object::GenerateShot(Vector3D pos, double theta_, double phi_) {
 	GenerateCube(1.0);
 	theta = theta_;
 	phi = phi_;
@@ -398,7 +398,7 @@ void Object::GenerateShot(Vector3D pos, float theta_, float phi_) {
 	RotateToHeading();
 }
 	
-void Object::GeneratePlayer(Vector3D pos, float theta_, float phi_,const unsigned short* texture, const int texWid, const int texHgt) {
+void Object::GeneratePlayer(Vector3D pos, double theta_, double phi_,const unsigned short* texture, const int texWid, const int texHgt) {
 	tex = texture;
 	texWidth = texWid;
 	texHeight = texHgt;
@@ -429,22 +429,22 @@ Point3D Object::GetPosition(void) const {
 	return position;
 }
 
-bool Object::SetBoudingEllipsoid(float x, float y, float z) {
+bool Object::SetBoudingEllipsoid(double x, double y, double z) {
 	rx = x;
 	ry = y;
 	rz = z;
 	return true;
 }
 
-float Object::GetRadiusX(void) const {
+double Object::GetRadiusX(void) const {
 	return rx;
 }
 
-float Object::GetRadiusY(void) const {
+double Object::GetRadiusY(void) const {
 	return ry;
 }
 
-float Object::GetRadiusZ(void) const {
+double Object::GetRadiusZ(void) const {
 	return rz;
 }
 
@@ -517,7 +517,7 @@ void Object::projectileInit(Vector3D head, Vector3D pos) {
 
 
 bool Object::CollidesWith(const Object& b) {
-	float distance = magnitude(b.GetPosition() - position);
+	double distance = magnitude(b.GetPosition() - position);
 	//if ( distance >= (radius + b.getradius()) )
 	//	return false;
 	if ( !updateList() )
@@ -533,7 +533,7 @@ bool Object::CollidesWith(const Object& b) {
 					 B = Vector3D(it->v[1].x,it->v[1].y,it->v[1].z),
 					 C = Vector3D(it->v[2].x,it->v[2].y,it->v[2].z);
 			Vector3D p,n1,n2,n3;
-			float t = (-it->normal * (p0 - A)) / (it->normal * (p1-p0));
+			double t = (-it->normal * (p0 - A)) / (it->normal * (p1-p0));
 			if ( t > 0 && t < 1 ) {
 				p = p0 + (p1-p0)*t;
 				n1 = normalize(Cross(A-B,p-B));
