+++ Engine/point.cpp
@@ -22,17 +22,17 @@ bool Point3D::operator!= (const Point3D& rhs) const
     return true;   
 }
 
-Point3D Point3D::operator* (const float rhs) const 
+Point3D Point3D::operator* (const double rhs) const 
 {
     return Point3D(x*rhs,y*rhs,z*rhs);//,w*rhs);   
 }
 
-Point3D Point3D::operator/ (const float rhs) const 
+Point3D Point3D::operator/ (const double rhs) const 
 {
     return Point3D(x/rhs,y/rhs,z/rhs,w/rhs);   
 }
 
-float Point3D::operator* (const Point3D& rhs) const 
+double Point3D::operator* (const Point3D& rhs) const 
 {
     return (x*rhs.x + y*rhs.y + z*rhs.z);// + w*rhs.w);
 }
@@ -57,17 +57,17 @@ Point2D& Point2D::operator= (const Point2D& rhs)
     return (*this);   
 }
 
-Point2D Point2D::operator* (const float rhs) const 
+Point2D Point2D::operator* (const double rhs) const 
 {
     return Point2D(x*rhs,y*rhs,w*rhs);   
 }
 
-Point2D Point2D::operator/ (const float rhs) const 
+Point2D Point2D::operator/ (const double rhs) const 
 {
     return Point2D(x/rhs,y/rhs,w/rhs);   
 }
 
-float Point2D::operator* (const Point2D& rhs) const 
+double Point2D::operator* (const Point2D& rhs) const 
 {
     return (x*rhs.x + y*rhs.y);
 }
